requirements.txt
langchain>=0.3.0
langchain-google-genai>=2.0.0
pyyaml>=6.0
